+++ crates/swc_diff/src/js_ast/ident.rs
