+++ crates/swc_diff/src/js_ast/stmt.rs
